Remove unnecessary context
