--- src/app/state/contexts.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const DatasetLoaderContext = React.createContext({});
